// src/app/generated/cite.api/model/teamMembership.ts
/*
 Copyright 2025 Carnegie Mellon University. All Rights Reserved.
 Released under a MIT (SEI)-style license. See LICENSE.md in the
 project root for license information.
*/

/**
 * Cite API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TeamRole } from './teamRole';


export interface TeamMembership { 
    id?: string;
    teamId?: string;
    userId?: string | null;
    roleId?: string | null;
    role?: TeamRole;
}